serial fix2

+++ Arduino/lib/Communication/SerialManager.h
@@ -82,7 +82,7 @@ public:
     }
 
     //for now only int are supported
-    int* getvar(unsigned int index)
+    int16_t* getvar(unsigned int index)
     {
         return internalRegister.getIncomingDataHeader(index);
     }
